Move remaining impl classes out of @vortex/fs into src/main
Addresses Erri's review feedback that @vortex/fs should be as close to
a type-only package as possible. Moves NodeFileSystemImpl,
WebFileSystemImpl, and PathResolverRegistryImpl into
src/main/src/filesystem/ and updates all consumers to use local imports.
